# Better BF16 support on AVX2